more variant software selection needles for BGO #771127
so what seems to be going on on the software selection screen
is some kind of GTK+ bug:

https://bugzilla.gnome.org/show_bug.cgi?id=771127

the radio button bullets don't always render correctly - I think
they're not always completing a transition they go through on
selection. I think they can get stuck in any state between 'grey
and small' and 'black and big', but for now these are the needles
I've managed to create from failures; we're missing a 'greyer
KDE' variant, but if that happens, we can add it. If the bug
gets fixed we should be able to drop all these.